requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
pypdf
python-multipart
black
fpdf2
requests 